+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Implement mapping profile details action menu. UIDEXP-75.
 * Fix spacing display of Transformation fields on mapping profile view page. UIDEXP-129.
 * Add job profiles deleting. UIDEXP-111.
+* Refactor to miragejs from bigtest/mirage. UIDEXP-137.
 
 ## [2.0.1](https://github.com/folio-org/ui-data-export/tree/v2.0.1) (2020-07-09)
 [Full Changelog](https://github.com/folio-org/ui-data-export/tree/v2.0.0...v2.0.1)